requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0
pytrends==4.9.2
supabase==2.27.1
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.7.1
google-generativeai==0.8.3
